requirements.txt
argparse
asgiref
celery
Django
django-admin
django-celery-beat
django-celery-results
django-contact-form
django-cors-headers
django-import-export
django-ratelimit
django-rosetta
djangorestframework
geoip2
gunicorn
openpyxl
pandas
psutil
psycopg2-binary
pyarrow
python-decouple
python-dotenv
python-slugify
pytz
redis
requests_cache
setuptools
sqlparse
watchdog
whitenoise
